Enable Compose and Treehouse usage in JS
This switches us to JetBrains' fork of AndroidX which has work-in-progress support for JS. Ideally their changes (or equivalent) will land upstream in AOSP at some point and we'll be able to switch back. For now, they're motivated to make this use-case work so they get our business.

The in-repo example already had a JS component which was only doing the display layer and communicating over websocket to a JVM backend running the compose layer. That server is removed in favor of running the compose layer directly in the browser solely for demonstration purposes. Having those layers exist separately is still a supported use-case, we just don't need to model it in the example.