Create event lambdas when property is set
This has two nice effects. The first is that it defers a bunch of lambda creation until needed. If you toggle an event property we will create two lambda instances instead of one, but that should be rare. The second advantage is that likely makes the widget interface viable as an indirection the generated compose code can talk through. This will enable compose to talk directly to widgets or produce diffs. Or perhaps produce schema instances for testing.