Enable Compose for native targets
This is possible now that we switched back to AOSP for Compose and support for native has landed upstream in AOSP. Once again we have to carry our own actuals for the runtime, but they are few (and sourced from the JetBrains fork).

For simplicity I have only enabled iOS, but I will follow up and enable all possible native targets for Treehouse artifacts.