Remove toolchain, set source/target compatibility
This has the same desired effect of the original change in that the artifacts are built with Java 8 bytecode, but does so in a way that allows using a modern JVM to build. With any luck, the new JVM will use less memory and make Windows builds less resource intensive.