Basic protocol/widget responsibility split
This is a first step toward the protocol being an optional step between the composable and the widget. It also has the side-effect of fixing the Objective-C/Swift view of the widget code through the removal of default function implementations on interfaces.

The composition can still only output protocol, but that will be changed in a follow-up.