Generate composition factory in regular compose module
This allows the factory to be properly typed, and the protocol factory for compose to simply be one such type which fullfils that contract. Currently, however, it is the ONLY supported implementation that works, despite a normal widget factory also able to be used. Its support will come in a follow-up.