Kotlin 1.6.10 and Compose upgrade
The compiler plugin artifacts (both shaded and unshaded) are now sourced from JetBrains. Additionally, the runtime is now built from JetBrains' AndroidX fork.

Compose 1.1.0 from Google was broken for JS and would not compile for native. The native fix was already in AOSP but the JS fix was not. This made bumping to the latest alpha or even HEAD of Compose from Google not an option to fix this problem.

We started on JetBrain's fork, and they run CI on JS and native, so now we are back to it. Thankfully some things have changed. For one, all of their artifacts redirect to Google's for the Android variant. This is the same thing we were already doing with our runtime.

However, while JetBrains' 1.1.0 does work for JS it does not work for native as that fix landed afterward. Thus, we are now on JetBrains' HEAD and Google's 1.2.0-alpha03. Hopefully JetBrains will release a 'stable' alpha of 1.2.0 that we can point to soon, but it doesn't really matter as we aren't interacting with other Compose libraries built for JS or native.

Oh and the runtime tests work now!