Upgrade to Kotlin 1.6.21 and Compose... something???
We can't wait any longer for JetBrains to release a version to Maven Central so we are jumping onto an alpha (published only to their Space instance). Note that the runtime SHA we are updating to does not immediately correspond to the compiler version from their space or the Android artifact from Google, but it should be fine. We won't be on such a bespoke setup for long.