Remove ability for schema to be multiplatform
This was added solely to for the schema-based testing strategy which needed to run on all platforms your Compose code could run on. With that removed, we can revert back to being JVM only as a simplification (as we require the use of JVM reflection to read the schema).

The future replacement of the testing strategy will support multiplatform and you can still pull multiplatform types into your schema as dependencies.