Stop running runtime tests
We only build the runtime because Jetbrains doesn't publish it with enough native targets that we need. However, they run the tests on all platforms they support, and we run tests in downstream modules on all targets we support. These tests are long and some are flaky (with little response to issues for fixing the flakes). Delete 'em!