Expose widget version to Compose
The value is only required when using the protocol to communicate between Compose and the Widgets where the version may skew. When using direct composition (with no protocol), the value is set to the max.

This WidgetVersion property is what we will lint against for compile-time checks along with a future annotation to indicate when a function requires a particular widget version.