Switch lambda schema type back to lambda
This more closely resembles the Compose code, and will allow the use of a receiver in the future to denote scope. Our current parsing mechanism does not provide access to the receiver in any meaningful way that we can differentiate from a normal lambda argument so its support is deferred.