Require only a Widget.Children to create a ProtocolDisplay
This simplifies setup because generally you have an existing container into which you want to place content and can wrap it in one of our children helpers. Internally, we synthesize a root widget in the same way the Treehouse code used to do.

Since we no longer need a widget, we no longer need to expose the public 'wrap' functions from the diff-consuming widget factory. These are now private and changed slightly to keep the 'when'-containing function body as small as possible.