Implement schema dependencies
A schema may now depend on others and use their widgets and layout modifiers.

To produce a schema as a library you need to have the schema itself as well as the generated artifacts for layout modifiers and widgets. You do not need to generate compose code or any protocol code.

To consume a schema, add it as a dependency to your `@Schema` annotation along with a tag (1-2000). The layout modifier and widget artifacts for your schema will need to depend on the associated artifacts from the dependency. The code
generation for the compose and protocol artifacts will automatically include all dependency types.