Remove the concept of clear
For Compose, this Applier function is only invoked when the Composition is going away and a new one is going to attach to an existing Applier. We never do that, thus, this function will never be invoked. Clear was used by Treehouse so a similar mechanism is introduced at that layer.