Move ID generation out of applier
An `IdAllocator` type creates IDs and is owned by the root factory. This allows IDs to be created synchronously at construction time.

Some other small changes that are tangentially related are included in this change:
- The type is not longer held in a backing field and instead each subtype has a property getter to return its value.
- Elements of the diff-producing widgets have been moved to the `DiffProducingWidget` interface such as ID, type, and the ability to send an event.

This is the first in many steps toward sharing an Applier implementation between protocol-based and non-protocol-based usage.