Reduce listener callbacks to one
This listener is was used for loading indicators and displaying animations on reloads. For loading indicators, display them upon initial creation of the TreehouseView subtype or when one is reused (such as in a data binding list view). For animating, the provided boolean differentiates first load from subsequent loads and it runs at the same time as before.

The view subtype parameter is not needed since you can capture the actual TreehouseView subtype when creating the lambda.