Use codegen to de-duplicate Compose helpers
This isn't great, but it's also not terrible. These never change so not really a big deal. The biggest downside is that the IDE does not run the task on sync and so you need to do a build before they resolve in the IDE. They will always link during compilation.