Make MeasureSpec a value type
This value is forced to be non-negative which does give us one free bit to manipulate, but there are no other easy bits to steal from a floating-point representation. Instead, use the two least-significant bits from the fractional part. This
will slightly alter values, but at precisions which are imperceptible to users.