Revert back to each schema generating its own composables
This ensures that a library can ship helper composables which depend on their own codegen.

Technical details of the change follows:

In order to support each schema's composables to call its factory functions there are now two interfaces generated.
The first is the traditional widget factory we've always had (although now without a marker interface), and the second is a provider of that widget factory called a provider (single property interface).
Every generated composable casts a provider class to its provider interface in order to access its factory and then calls the specific widget function.

    @Composable
    fun Row(..) {
      RedwoodComposeNode<RedwoodLayoutWidgetFactoryProvider<*>, Row<*>>(
        factory = { it.RedwoodLayout.Row() },

This is the mechanism which allows multiple schemas to compose.
The class which implements the provider interface downstream can implement many of these interfaces to support schema dependencies.

Each schema generates a helper class for providing the factory and all of its dependant factories.

    class EmojiSearchWidgetFactories<W : Any>(
      override val EmojiSearch: EmojiSearchWidgetFactory<W>,
      override val RedwoodLayout: RedwoodLayoutWidgetFactory<W>,
      override val RedwoodTreehouseLazyLayout: RedwoodTreehouseLazyLayoutWidgetFactory<W>,
    ) : EmojiSearchWidgetFactoryProvider<W>

This is what is constructed by calling code and provided to the initialization function.