Generate less serialization code
No more empty object to hold a function which just returns a LayoutModifierElement with an integer. That is inlined to the callsite.

No more secondary constructor to unwrap the real LayoutModifier into its component values which is only called from a companion function. That unpacking is done directly in the companion function now.