Respect event type nullability
If an event is declared as non-nullable then it is required and non-null. If an event is declared as nullable then it is still required but can be set to null. And, of course, nullable event can have a default of null.

This brings it directly in line with property behavior. I'm not sure why we had the weird nullability behavior from before.