Build a new AppLifecycle abstraction to host the frame clock
Previously we sent a frame 60 times per second, whether one
was requested or not.

With this change we establish bidirectional communication
between the guest and the host, and use that to make a frame
clock that only emits frames when they're needed.

This has significant performance benefits, particularly when
the application is idle.