Assign a string key to every single item in a LazyColumn
This fixes issues with items being recreated repeatedly.

I don't like that we need to create lots of strings with this,
but it's a simple fix that's likely to prove useful with other
features like keeping scroll position.

I've only honored this for Android views. For follow-up it's
necessary to honor it for Compose and iOS.