Remove test for Compose UI packaging
This test was relevant when we built our own Compose runtime and had to do special build things to ensure Gradle merged it with the AndroidX Compose runtime rather than double-packaging it. We now rely on the JetBrains Compose runtime (which also does this trick), but it's no longer our problem and therefore we do not need to test the behavior. Also, our Compose UI samples still implicitly test this behavior.