Change our tree to be custom nodes instead of raw widgets (#1017)
We already were faking half of the tree with the synthetic children widgets before. Now we keep children nodes which do the same thing as before but are joined by widget nodes which wrap a widget and provide access to the parent children. This is used to notify the parent children when any child's layout modifiers have changed.