Rewrite the protocol to be more precise
The new protocol has two features not present in the old one:
- Creation of a node is separate from attaching it to a parent node's children.
- Mutations of properties, layout modifiers, or the tree structure are recorded precisely in the order they occur by Compose

This change was primarily motivated by the desire to have nodes on the widget-side be created, have all their initial properties set, and then be attached to their parents.

As part of this change there is also a slight change in the order of widget attachment in multi-level branches of the tree. Prior to this change, given

    Row {
      Column {
        Text("Hi")

We would see `Column` created and attached to `Row`, and then `Text` created and attached to `Column`. With this change, however, `Text` is attached to `Column` before `Column` is attached to `Row`.

Basically, when a new branch is attached you can assume it is fully formed structurally (as well as having initial properties set).