Fix JS build by specifying unique module names
By default these JS modules inherit the project name which was either 'browser' or 'presenter-treehouse', except those names are used twice because we have multiple samples with the same folder structure. When these modules were installed into a repo-local folder of all modules, the duplication would cause the second task to overwrite the first. As of Gradle 8, this is a built-time error.

Specifying unique module names ensures these write to separate folders and thus are able to be built without stomping on each other.