Eliminate properties used for publishing artifacts (#1202)
The project name is now used as the artifactId automatically. These all matched the explicitly-specified ones from before.

I have eliminated the name which is now also set to the project name. Prior to this it was set to a more human-readable version of the artifactId.

This was tested by installing all artifacts to a local Maven repo both before and after this change and then diffing the directories. The only changes to the poms is the name change. For example:

    diff -r build/localMavenOld/app/cash/redwood/redwood-treehouse-js/0.5.0/redwood-treehouse-js-0.5.0.pom build/localMaven/app/cash/redwood/redwood-treehouse-js/0.5.0/redwood-treehouse-js-0.5.0.pom
    13c13
    <   <name>Redwood Treehouse</name>
    ---
    >   <name>redwood-treehouse</name>
    15a16