Update UITableView to track changes precisely
Previously we called reloadData() whenever anything changed
in the underlying model.

With this change we fire precise events to the UI when inserts
and removes happen in the source LazyList. This attempts to
coalesce changes to itemsBefore() and itemsAfter() with
adjacent changes and the beginning and end of the items list.
To implement this the changes are buffered into an intermediate
model before they are applied.

When a cell changes from being a placeholder to a loaded
item, or the opposite, this does a cell content change
without firing an event through the UITableView. To
implement this we must track which cells are currently
displaying placeholders. This PR includes a new datastructure,
SparseList, to implement this tracking.

This PR also changes the number of placeholders that
guest code offers to host code. I found experimentally
that 20 placeholders was not enough for UITableView.

This PR changes LazyList from UICollectionView to
UITableView. This change was potentially unnecessary,
though UITableView has fewer features that we don't
need.