Remove exotic Modifier APIs, fix normalization
As we consider removing/reimagining Modifier as a concept in our composables, remove more of their exotic API which is focused on usage as interceptor-like structures (from Compose UI) rather than as key/value structures (from Redwood).

Added tests for remaining API behavior, including that of an inconsistency in how combined modifiers behave. The CombinedModifier type retains association on how it was assembled despite being otherwise a list-like structure. So ((a then b) then c) did not equal (a then (b then c)) in old code. Now, if an CombinedModifier is being added to another, its elements are unpacked forming a normal form. Additionally, since nothing references this type or needs to use it directly, remove it from the public API.