Eliminate allocation and boxing in onScrolled callback. (#1609)
This callback is invoked on every frame when scrolling and therefore should not produce a lot of allocations. Previously it allocated multiple iterators/sequences and was subject to integer boxing and unboxing. Unrolling the functional operations into a traditional loop keeps data in primitives and never allocates an iterator or sequence.