Support animated programmatic scrolls
In emoji search when the search terms is changed we scroll
to the top of the list. This means we need to treat
programmatic scrolls as something that can happen repeatedly,
not just when we restore a lazy list.

This moves the feature that prevents programmatic scrolls
from colliding with user scrolls from host code to guest code.
The benefit is it no longer risks getting the loaded window
out of sync with the true scroll window.

This also implements all scroll behavior for iOS.