Move protocol modifier deserialization into factory (#1640)
This is in preparation of adding index tracking to the protocol nodes to match support in the native widget nodes. ProtocolNode became an abstract class because of this, and common functionality to all subtypes moved into the class.

Additionally, create a split between the protocol factory which users need to reference and the generated protocol factory APIs that are unstable and used internally. This simplifies how these types and API signatures which use them look to users.