CodeListener.onUncaughtException
This new API is called on host views when the guest code fails
with an uncaught exception.

This wires in a CoroutineExceptionHandler in the host code
that cancels the current Zipline instance and updates all
UIs its currently serving, if any.

There is not yet any support for exceptions that occur in
async code, or any mechanism to restart the Zipline instance
after a crash.