Change TreehouseUIKitView's size as necessary
If the content grows or shrinks, this reflects that.

I really didn't want to delegate to UIStackView here because this
layout isn't a column at all. But after fighting it for hours and
hours I gave up because I couldn't figure out how to react to my
child views' size changes. I suspect the fix may be to add a listener
on the bounds of the child views; nothing in the docs recommends
this however. I'd love to know how UIStackView implements it!