Change TreehouseUIKitView's size as necessary (#1858)
* Change TreehouseUIKitView's size as necessary

If the content grows or shrinks, this reflects that.

I really didn't want to delegate to UIStackView here because this
layout isn't a column at all. But after fighting it for hours and
hours I gave up because I couldn't figure out how to react to my
child views' size changes. I suspect the fix may be to add a listener
on the bounds of the child views; nothing in the docs recommends
this however. I'd love to know how UIStackView implements it!

* Spotless

* Update redwood-treehouse-host/src/iosTest/kotlin/app/cash/redwood/treehouse/TreehouseUIViewLayoutTest.kt

Co-authored-by: Jake Wharton <jw@squareup.com>

* Update redwood-treehouse-host/src/iosTest/kotlin/app/cash/redwood/treehouse/TreehouseUIViewLayoutTest.kt

Co-authored-by: Jake Wharton <jw@squareup.com>

* Update redwood-treehouse-host/src/iosTest/kotlin/app/cash/redwood/treehouse/TreehouseUIViewLayoutTest.kt

Co-authored-by: Jake Wharton <jw@squareup.com>

* Update CHANGELOG

---------

Co-authored-by: Jake Wharton <jw@squareup.com>