Create RedwoodVersion type for sharing across host-guest bridge (#1929)
In order to work around bugs in the future or change the protocol in a backwards-incompatible way, the version gives the other side of the bridge the ability to change itself to remain compatible.

For now, we only have the host version hooked up to the guest side. The guest version being exposed to the host will come in a follow-up.