Don't crash when loaded rows are removed.
We had a bug where we asynchronously detached a LazyList item
from its parent, but synchronously recycled it. We didn't see
this occur in practice very much because the bug only triggered
when the removed row was in the middle of the loaded range. Rows
at the top or bottom of the loaded range were synchronously
turned into placeholders when they are removed from the loaded
range.

One tough limitation of this approach is the removed items
lose their content immediately, which may interfere with
delete animations. This isn't an urgent problem for us but
it suggests that we should explore an API to put items into
the pool asynchronously and not the moment they're logically
removed.