Unify multiplatform targets into named groups
This gives us more control over what targets are built for which parts of the codebase based on their type.

The big change is the drop of MacOS targets which were wholly unused. They will likely return at some point, along with new targets for common modules as well as for Compose UI, and this new grouping will make that easy to do.

Before:

    $ gw build --dry-run | grep -c SKIPPED
    7469

After:

    $ gw build --dry-run | grep -c SKIPPED
    6405