New Widget.detach() is a user function called at detach time (#2083)
* New Widget.detach() is a user function called at detach time

This is invoked when a view will never be operated by the
presenter. It usually happens when a view is removed from
the view tree, but may also happen later if the view or one
of its parent views is pooled.

Implementations should do what they can to make Kotlin objects
eligible for immediate garbage collection. In particular a
useful thing to do is to remove references from reference-counted
Swift objects to Kotlin objects. That way the GC can reclaim
the memory sooner, and we may reduce the chance of encountering
KT-68769.

https://youtrack.jetbrains.com/issue/KT-68769/EXCBADACCESS-Attempted-to-dereference-garbage-pointer-0x28

* Address PR feedback. apiDump.