Fix a host crash in LazyListUpdateProcessor
I introduced a bug in the mechanism that attempts to replace
animated inserts and removes with update-in-place transitions.

The crash occurs when the before window shrinks while the total
number of elements is growing. The previous code naively
assumed the before window shrinking would always be a scroll,
but it can coincide with a content change.

Closes: https://github.com/cashapp/redwood/issues/2172