Optimize Modifier serialization in guest code
I learned that encoding a type as a JsonElement and then
encoding that as JSON is significantly less efficient than
going directly from the type to JSON. This is due to some
inefficient code in the internals of kotlinx.serialization.

Skipping this intermediate step is useful on its own, but it
also turns out to be significantly more efficient in practice.