Don't pass UIViewNoIntrinsicMetric to sizeThatFits
This is a reasonable value to return from a measurement but
I don't think it makes sense as an argument when making a
new measurement.

I've observed UIButton returning the wrong size when passed
this value, and the correct size when passed 0.0. Unfortunately
I'm unable to reproduce that behavior without a lot of code
set up.