Do not store the WidgetTag + base impl of visitIds
Storing the WidgetTag is a waste of space. We rarely use it, and even then it's only an int so we can just codegen returning it directly.

Do not generate implementations of visitIds for nodes with no children. Have the base class invoke the lambda with the current node ID by default. Only nodes for widigets with children will override and do codegen.