Flatten RedwoodView and RedwoodView.Root
This is a step in simplifying the complex interaction between these
two classes. It also removes an unnecessary layer from the view
tree.

We also had a potential bug on hot reloads, where the previous content
was detached but not necessarily removed. With this change we clear
the child views by default on a hot reload.