Implement ViewBox with manual measurement
Previously ViewBox delegated to FrameLayout. This worked okay
but it was potentially difficult to track behaviors between
View and UIView.

This implementation is based on the UIViewBox class implementation.

This fixes a bug where layouts incorrectly didn't handle changes
when modifiers were removed.

This changes how Box's margins are implemented. Previously the
margins were external, which required the hosting layout to
honor them as layout constraints. That was not supported when
the enclosing layout was a Row or Column.