Host-side API for view-owned widgets
This introduces a new mutable property, windowInsets, that is propagated to the composition where it can be consumed.

We put the burden on the user of RedwoodView to set this value to something non-zero if they have insets to be consumed. I've added an accelerator on Android for this,

RedwoodLayout.consumeWindowInsets(mask: Int)